style(server):refactor sign up feature
