+++ server/app.js
@@ -1,7 +1,6 @@
 import express from 'express';
 import bodyParser from 'body-parser';
-import data from './db/data';
-import user from './src/models/user.model';
+import UserController from './src/controllers/user.controller';
 
 // set up the express app
 const app = express();
@@ -12,140 +11,7 @@ app.use(bodyParser.urlencoded({ extended: false }));
 
 const baseUrl = '/api/v1/';
 
-/**
- * Register a user
- * @param {object} req
- * @param {object} res
- */
-app.post(`${baseUrl}auth/signup`, (req, res) => {
-  if (!req.body.email) {
-    return res.status(400).send({
-      success: false,
-      message: 'email is required',
-    });
-  }
-
-  const email = /.+@.+\..+/;
-  if (!email.test(req.body.email)) {
-    return res.status(400).send({
-      success: false,
-      message: 'email is invalid',
-    });
-  }
-
-  if (req.body.email === data.email) {
-    console.log(data.email)
-    return res.status(409).send({
-      success: false,
-      message: 'email exists already',
-
-    });
-  }
-
-  if (!req.body.first_name || !req.body.last_name) {
-    return res.status(400).send({
-      status: error,
-      error: 'first name or last name is required',
-    });
-  }
-
-  const letters = /^[0-9a-zA-Z]+$/; // names must only contain alphanumeric char
-  if (!letters.test(req.body.first_name) || !letters.test(req.body.last_name)) {
-    return res.status(400).send({
-      status: error,
-      error: 'first name or last name contains invalid charactor(s)',
-    });
-  }
-
-  if (req.body.first_name.length < 2 || req.body.last_name.length < 2) {
-    return res.status(400).send({
-      status: error,
-      error: 'first name or last name must contain at 2 charactors',
-    });
-  }
-
-  if (!req.body.password) {
-    return res.status(400).send({
-      status: error,
-      error: 'passord is required',
-    });
-  }
-
-  if (req.body.password.length < 8) {
-    return res.status(400).send({
-      status: error,
-      error: 'Minimum passord length is 8',
-    });
-  }
-
-  const abc = /^[a-zA-Z]$/;
-  if (abc.test(req.body.password)) {
-    return res.status(400).send({
-      status: error,
-      error: 'invalid password. must contain at least 1 special charactor',
-    });
-  }
-
-  const user = {
-    id: data.length + 1,
-    email: 'sarah@gmail.com',
-    fist_name: 'Sarah',
-    last_name: 'Lifaefi Masika',
-    password: 'usr$_18@',
-    is_admin: false,
-  };
-
-  data.push(user);
-  return res.status(201).send({
-    status: 'success',
-    message: 'user registered successfully',
-    data,
-  });
-});
-
-/**
- * Sign in a user
- * @param {object} req
- * @param {object} res
- */
-app.post(`${baseUrl}auth/signin`, (req, res) => {
-  if (!req.body.email) {
-    return res.status(400).send({
-      status: 'error',
-      error: 'email is required',
-    });
-  }
-
-  const email = /.+@.+\..+/;
-  if (!email.test(req.body.email)) {
-    return res.status(400).send({
-      status: 'error',
-      error: 'email is invalid',
-    });
-  }
-
-  if (!req.body.password) {
-    return res.status(400).send({
-      status: 'error',
-      error: 'passord is required',
-    });
-  }
-
-  const foundUser = user.findUserByEmail(req.body.email, req.body.password);
-
-  if (!foundUser) {
-    return res.status(404).send({
-      status: 'error',
-      error: 'user with login/pass does not exist, register first',
-    });
-  }
-
-  if (foundUser) {
-    return res.status(200).send({
-      status: 'success',
-      data: foundUser,
-    });
-  }
-});
+app.post(`${baseUrl}auth/signup`, UserController.signUpUser);
+app.post(`${baseUrl}auth/signin`, UserController.logInUser);
 
 export default app;
