+++ server/src/db/data.js
@@ -1,0 +1,22 @@
+const data = [
+  {
+    id: 1,
+    email: 'sarah@gmail.com',
+    fist_name: 'Sarah',
+    last_name: 'Lifaefi Masika',
+    password: 'usr$_18@',
+    is_admin: false,
+
+  },
+  {
+    id: 2,
+    email: 'dklf@gmail.com',
+    fist_name: 'Kelavos',
+    last_name: 'Lav',
+    password: 'usr$_18@',
+    is_admin: false,
+
+  },
+];
+
+export default data;
